[SDCC] Fixed sdasz80 assembler to allow export of 32-bits virtual addresses. [ASM] Added macro for optimized waiting time. [Sample] Fixed localize sample data (Japanese). Updated NEO mapper sample with more test case.

+++ projects/targets/s_target.c
@@ -559,7 +559,7 @@ void DiplayInfo()
 		
 		Print_DrawFormat("\nSize:   %s", GetROMSize(ROM_SIZE));
 		#if (ROM_MAPPER != ROM_PLAIN)
-			Print_DrawFormat(" (%i)", ROM_SEGMENTS);
+			Print_DrawFormat(" (%i)", (i16)ROM_SEGMENTS);
 		#endif
 	#endif
 	Print_DrawFormat("\nAddr:   %4x~%4x", Sys_GetFirstAddr(), Sys_GetLastAddr());
